feat: simulator - final
