+++ med-equip-system-back/src/main/java/medequipsystem/config/WebSecurityConfig.java
@@ -85,6 +85,7 @@ public class WebSecurityConfig {
                                         .antMatchers(HttpMethod.GET, "/api/auth/confirm")
                                         .antMatchers(HttpMethod.GET, "/api/companies/all")
                                         .antMatchers(HttpMethod.GET, "/api/companies/{id}")
+                                        .antMatchers(HttpMethod.GET, "/api/startsending")
 
                 .antMatchers(HttpMethod.GET, "/", "/webjars/**", "/*.html", "favicon.ico",
                         "/**/*.html", "/**/*.css", "/**/*.js");
