+++ med-equip-system-front/src/app/company-admin/location-simulator/location-simulator.component.html
@@ -1,8 +1,7 @@
 <div class="center"> 
     <app-simulator-map [markerLatitude]="currentLocation.latitude" [markerLongitude]="currentLocation.longitude"></app-simulator-map>
+    <p> Vehicle coordinates: {{currentLocation.latitude}} , {{currentLocation.longitude}} </p>
     <button class="primary" (click)="startDelivery()">Start delivery</button>
-    <p> {{currentLocation.latitude}} , {{currentLocation.longitude}} </p>
-    <button class="primary" (click)="startSendingCoordinates()">Start sending coordinates</button>
 
 </div>
